Plot distances

This function will plot a barplot using the output form either within.dist or between.dist. Many of the same parameters as barplot() can be used with this function.

plot.dists(list, space = NULL, col = c(), ylim = c(0, 1), ylab = "Average Distance", xlab = "", main = c())

Arguments

list
a list of vectors containing the distances between or within samples

space
the amount of space (as a fraction of the average bar width) left before each bar. May be given as a single number or one number per bar. If height is a matrix and beside is TRUE, space may be specified by two numbers, where the first is the space between bars in the same group, and the second the space between the groups. If not given explicitly, it defaults to c(0,1) if height is a matrix and beside is TRUE, and to 0.2 otherwise.

col
a vector of colors for the bars or bar components. By default, grey is used if height is a vector, and a gamma-corrected grey palette if height is a matrix.

ylim
limits for the y axis.

ylab
a label for the y axis.

xlab
a label for the x axis.

main
overall title for the plot

Details

The only necessary function is 'list'. If no 'col' is specified, then the function will randomly generate colors. For the rest of the parameters the defaults work well for quickly visualizing distances.

Value

References

Note

See also

Examples

# Plot output from other function using random colors and default parameters
#plot.dists(btwn.dists)	

# Specifying colors and main title
#plot.dists(btwn.dists, col=c('blue','red','green'), main='Title of Chart')

## The function is currently defined as
function (list, space = NULL, col = c(), ylim = c(0, 1), ylab = "Average Distance",
    xlab = "", main = c())
{
    avgs <- c()
    for (i in 1:length(list)) {
        avgs[i] <- mean(list[[i]])
    }
    if (length(col) == 0) {
        col <- rgb(runif(length(list)), runif(length(list)),
            runif(length(list)))
    }
    xcor <- barplot(avgs, names.arg = names(list), space = space,
        ylim = ylim, col = col, xlab = xlab, ylab = ylab, main = main)
    for (i in 1:length(list)) {
        arrows(x0 = xcor[i], x1 = xcor[i], y0 = mean(list[[i]]),
            y1 = mean(list[[i]]) + sd(list[[i]]), angle = 90)
    }
  }

#> function (list, space = NULL, col = c(), ylim = c(0, 1), ylab = "Average Distance", 
#>     xlab = "", main = c()) 
#> {
#>     avgs <- c()
#>     for (i in 1:length(list)) {
#>         avgs[i] <- mean(list[[i]])
#>     }
#>     if (length(col) == 0) {
#>         col <- rgb(runif(length(list)), runif(length(list)), 
#>             runif(length(list)))
#>     }
#>     xcor <- barplot(avgs, names.arg = names(list), space = space, 
#>         ylim = ylim, col = col, xlab = xlab, ylab = ylab, main = main)
#>     for (i in 1:length(list)) {
#>         arrows(x0 = xcor[i], x1 = xcor[i], y0 = mean(list[[i]]), 
#>             y1 = mean(list[[i]]) + sd(list[[i]]), angle = 90)
#>     }
#>   }
#> <environment: 0x7fad4d6f7e50>