Not Sure Yet

preforms wilcox test on the given groups of a phylip formatted distance matrix. .

wilcox.dist(dist,design,groups,subgroups=FALSE)

Arguments

dist
Either a file containing a phylip formatted distance matrix with ThetaYC, UniFrac, or some other distance metric between samples or a variable representing the distance matrix that was imported using the read.dist function.

design
a '.design' file, i.e. a file containing two columns separated by a tab, with the sample names in the left column and the name of their group in the right column.

groups
a vector of strings containing the names of the groups to be analyzed. There is no limit to the number of groups that can be analyzed. The parameter 'all' can be used to analyze the relationships between all groups.

subgroups
a logical value determining whether to use the group column or the subgroup column in the analysis (FALSE is groups TRUE is subgroups). (which subgroups to use are determined by the groups vector.)

between
A logical value If FALSE wilcox.dist takes the distances of all the points of each group relative to all the other points of the same group then prints a wilcox rank sum test comparing every 2 groups. If TRUE wilcox.dist the distances of all the points of each group relative to all the other points of another group for example if there are 3 groups, 'A', 'B' and 'C', wilcox.dist outputs the wilcox rank-sum test of AB vs AC, AB vs BC and AC vs BC.

Details

Value

References

Note

See also

Examples